fix drag and drop, rename content-renderer
