+++ src/default-inner-content-renderer.js
